scraper/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.18.1
psycopg2
python-dotenv==1.0.1
pandas==2.2.0
lxml==5.1.0
fake-useragent==1.5.1